requirements.txt
aiofiles==0.6.0
amqp==1.4.9
anyjson==0.3.3
autopep8==1.5.4
billiard==3.3.0.23
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dataclasses==0.8
fastapi==0.63.0
h11==0.12.0
importlib-metadata==3.3.0
kombu==3.0.37
Pillow==8.1.0
prompt-toolkit==3.0.9
pycodestyle==2.6.0
pydantic==1.7.3
python-multipart==0.0.5
pytz==2020.5
six==1.15.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
uvicorn==0.13.3
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
